gdbus: Improve registered paths activation on daemon side
This changes the way registered paths (i.e. interface skeleton exports)
are activated on daemon side. We're keeping a list of connections on which every
registered path is exported so that if connection is closed or the path is
unregistered we always free the right interface skeletons.

We also keep track of active client (peer-to-peer) connections so that if a new
path is registered, it's automatically exported to all active connections. Needed
since client connections are usually shared.

It also has a nice side-effect of reacting to dirty disconnects (e.g. the client
dies).

This whole registration machinery needs unique object paths (for the hash table)
but that's not a problem in the current state.