gdbus: Flush the async metadata calls
Otherwise it would not be guaranteed that the messages got actually
sent if a program exits afterwards.

Synchronously flushing queued messages on a shared GDBusConnection
is quite risky as the queue could be rather long and we would block,
however the affected calls are not that common in normal use.